refactor lilconfig to js
