+++ package.json
@@ -2,15 +2,11 @@
     "name": "lilconfig",
     "version": "3.1.0",
     "description": "A zero-dependency alternative to cosmiconfig",
-    "main": "dist/index.js",
-    "types": "dist/index.d.ts",
+    "main": "src/index.js",
+    "types": "src/index.d.ts",
     "scripts": {
-        "prebuild": "npm run clean",
-        "build": "tsc --declaration",
-        "postbuild": "du -h ./dist/*",
-        "clean": "rm -rf ./dist",
         "test": "NODE_OPTIONS=--experimental-vm-modules ./node_modules/.bin/jest --coverage",
-        "lint": "eslint ./src/*.ts"
+        "lint": "eslint ./src/*.js"
     },
     "keywords": [
         "cosmiconfig",
